Fixes the race conditions identified on github issue #8:
There are 2 that I know of:
1. Logger If the operational program is in the middle of filling in an entry to the logger shared memory and the control program copies it, the logging information may be only partially updated. When running address sanitizer, I am seeing this blow up sometimes.
2. Deferred Serialization If the operational program does a "write" to a statistic requiring deferred serialization, this may race with the periodic "serialization" thread that also runs in the operational program.

Case 1 will require a shared semaphore (between the control and operational programs). Care has to be take so that the operation program NEVER blocks -- it would probably just try to get the lock when it is has logging information to write and if the lock is not available it will just drop the log entry.

Case 2 will require a simple mutex in the operational program. Again, the "write" path should never block.